+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 Changelog
 
 ### unreleased
+
+## 0.2
++ Added log level to NSURLSession
++ Added consistent logging pattern 
+
+## 0.1
 + Mock single requests to NSURLConnection
 + Add delay to mocked responses on NSURLConnection
 + Mock NSURLSession simple init methods